Initial Launching Version
This version of the updated 00_cb_types.txt will launch but may
not be perfect. There are references to remove and I want to
see if we can outright remove (or comment out) the references
to muslims, dharmic religion, china, Jade Dragon tributary types,
and so on. It launched on my machine and I tried a few rulers, they
seemed to have the correct CBs, none were broken missing or added.
Let me know if you find something that shouldn't be there