Makes Rastafarian less progresive
In the  Rastafarian religion woman are not afforded the same rights as other faiths. Some of the rules are as follows:
Women are regarded as subordinate to men
Women are regarded as housekeepers and child bearers
Women cannot be leaders
Men are the spiritual head of the family
These changes make the Rastafarian faith even more unique compared to the other afro syncretic faiths.