Clampet -> Mullins
Mullins is a very common name in Southwest Virginia, which is where this dynasty resides in After The End. It's also a very common Melungeon name. I'm also changing they dynasty culture to Melungeon, since every member of the dynasty that exists at the start of the mod is Melungeon.