Modifications of Starting Scenario in Yukon, Stikine and Alexander
New vassals and characters are added to the duchies of Yukon, Stikine and Alexander in the far northwest of the map.

Rich McIntyre is joined by younger brother and Vassal, Jack, who rules the county of Stikine. Between their personal demesnes is the elderly soon-to-be-widower Dalton Horwood. Dalton, has who gained title over this land by calling in a favour earned over two generations of dutiful service to the McIntyres and their father, now finds himself heirless, and desperate to carry on his ancient and storied line.

Aurore Godin, whose house displaced the McIntyres from Whitehorse a century ago, is the wife of her Tlingit liege, Yeilgooxu Xunaa. She had sons from a previous marriage to secure House Godin's claim to Whitehorse, but a fatal accident has recently befallen her youngest, and her husband's house now stands to gain from any misfortune that may befall her eldest. Should her line die out, the Tlingit are poised to inherit one of the great centers of Yukoner culture.

Elias Charlie is a legendary family man, blessed with children in abundance, putting him in an advantageous position for marital alliances with his Tlingit overlord, and many foreigners who might be in search of aliances should his ambitions grow beyond the limits of Kluane.

Walter LeConte, better known as 'The Black' is a corsair of legend across the Inside Passage. In the chaos following the overthrow of the Xunaas and the Haida conquest, he was able to leverage his martial skills, ruthlessness and willingness to convert to the Haida faith to secure for himself a position as a vassal of the Tlagaang. Now flying Raven's banner, this notorious pirate has a new opportunity to plunder his foes and neighbours like never before... or perhaps settle down and go legitimate, as much as a salt of his likes can.

There are also minor corrections outside of the discussed area. Character 1200001 is renamed from Tsaa Hamjuuwee to Tsaak, so that he does not share a name with his own brother, also named Tsaa Hamjuuwee, king of the Januug Tlagaang. The three daughters of Drake Messier have been corrected to have his dynasty and culture.